+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Development
 
+### Feature
+- Improve restapi look and add impressum
+
 ## 0.107.0 - 2025-03-25
 
 ### Refactor